+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/mixin/HuffConstantReferenceMixin.kt
@@ -1,13 +1,13 @@
 package com.github.com.cakevm.intellij_huff_plugin.language.psi.mixin
 
-import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffMacroConstantReference
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffConstantReference
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.impl.HuffNamedElementImpl
-import com.github.com.cakevm.intellij_huff_plugin.language.reference.HuffMacroConstantReferenceReference
+import com.github.com.cakevm.intellij_huff_plugin.language.reference.HuffConstantReferenceImpl
 import com.github.com.cakevm.intellij_huff_plugin.language.reference.base.HuffReference
 import com.intellij.lang.ASTNode
 import com.intellij.psi.PsiElement
 
-abstract class HuffMacroConstantReferenceMixin(node: ASTNode) : HuffNamedElementImpl(node), HuffMacroConstantReference {
+abstract class HuffConstantReferenceMixin(node: ASTNode) : HuffNamedElementImpl(node), HuffConstantReference {
 
   override val referenceNameElement: PsiElement
     get() = this.identifier
@@ -17,5 +17,5 @@ abstract class HuffMacroConstantReferenceMixin(node: ASTNode) : HuffNamedElement
 
   override fun getName(): String? = referenceName
 
-  override fun getReference(): HuffReference = HuffMacroConstantReferenceReference(this as HuffMacroConstantReference)
+  override fun getReference(): HuffReference = HuffConstantReferenceImpl(this as HuffConstantReference)
 }