+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/psi/stub/impl/HuffFileStub.kt
@@ -12,7 +12,7 @@ class HuffFileStub(file: HuffFile?) : PsiFileStubImpl<HuffFile>(file) {
 
   object Type : IStubFileElementType<HuffFileStub>(HuffLanguage.INSTANCE) {
     // Update this version number if the stub structure changes
-    override fun getStubVersion() = 7
+    override fun getStubVersion() = 8
 
     override fun getBuilder(): StubBuilder =
       object : DefaultStubBuilder() {