+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/reference/HuffConstantReferenceImpl.kt
@@ -1,14 +1,13 @@
 package com.github.com.cakevm.intellij_huff_plugin.language.reference
 
-import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffConstantReferenceExpression
+import com.github.com.cakevm.intellij_huff_plugin.language.psi.HuffConstantReference
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.element.HuffNamedElement
 import com.github.com.cakevm.intellij_huff_plugin.language.psi.parentRelativeRange
 import com.github.com.cakevm.intellij_huff_plugin.language.reference.base.HuffReference
 import com.github.com.cakevm.intellij_huff_plugin.language.reference.base.HuffReferenceBase
 import com.intellij.openapi.util.TextRange
 
-class HuffConstantReferenceExpressionReference(element: HuffConstantReferenceExpression) :
-  HuffReferenceBase<HuffConstantReferenceExpression>(element), HuffReference {
+class HuffConstantReferenceImpl(element: HuffConstantReference) : HuffReferenceBase<HuffConstantReference>(element), HuffReference {
 
   override fun calculateDefaultRangeInElement(): TextRange {
     return element.referenceNameElement.parentRelativeRange