+++ src/main/kotlin/com/github/com/cakevm/intellij_huff_plugin/language/reference/HuffResolver.kt
@@ -25,10 +25,7 @@ object HuffResolver {
           is HuffMacroLabelGoTo -> {
             resolveUsingImports(HuffMacroLabel::class.java, element, element.containingFile)
           }
-          is HuffMacroConstantReference -> {
-            resolveUsingImports(HuffConstantDefinition::class.java, element, element.containingFile)
-          }
-          is HuffConstantReferenceExpression -> {
+          is HuffConstantReference -> {
             resolveUsingImports(HuffConstantDefinition::class.java, element, element.containingFile)
           }
           is HuffBuiltinFnFuncSigCall -> {